clases dao added

+++ src/main/java/com/tecnm/you2be/models/UsuarioBuyVideo.java
@@ -3,13 +3,15 @@ package com.tecnm.you2be.models;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @RequiredArgsConstructor
 public class UsuarioBuyVideo {
 
     private int idCompra;
 
-    private double descuento;
+    private BigDecimal descuento;
 
     private int idUsuario;
